chore(NA): move grokdebugger plugin test fixtures out of __tests__ folder
